+++ app/src/main/java/com/brewthings/app/ui/screens/pill/GraphScreenState.kt
@@ -1,6 +1,6 @@
 package com.brewthings.app.ui.screens.pill
 
-import com.brewthings.app.ui.screens.pill.data.DataType
+import com.brewthings.app.ui.screens.pill.graph.DataType
 import com.brewthings.app.ui.screens.pill.graph.GraphState
 import com.brewthings.app.ui.screens.pill.insights.InsightsState
