--- app/src/main/java/com/brewthings/app/ui/screens/pill/data/DataType.kt
@@ -1,7 +1,0 @@
-package com.brewthings.app.ui.screens.pill.data
-
-enum class DataType {
-    GRAVITY,
-    TEMPERATURE,
-    BATTERY
-}
